+++ src/lib.rs
@@ -167,6 +167,7 @@
 //! | `codes_2d`    | Print 2D codes (QR Code, PDF417, GS1 DataBar, DataMatrix, Aztec, etc.) |   ✅    |
 //! | `graphics`    | Print raster images                                                    |   ❌    |
 //! | `usb`         | Enable USB feature                                                     |   ❌    |
+//! | `native_usb`  | Enable native USB feature                                              |   ❌    |
 //! | `hidapi`      | Enable HidApi feature                                                  |   ❌    |
 //! | `serial_port` | Enable Serial port feature                                             |   ❌    |
 //! | `full`        | Enable all features                                                    |   ❌    |